+++ js/components/listItem.jsx
@@ -2,8 +2,7 @@ var React = require ('react');
 var ReactDOM = require ('react-dom');
 var Backbone = require ('backbone');
 var BackboneParse = require ('../../backbone-parse.js');
-var App = require ('./mainApp.jsx');
-var Todo = require ('./listToDo.jsx');
+
 
 
 var ListItem = Backbone.Model.extend({
@@ -25,6 +24,7 @@ var ListItem = Backbone.Model.extend({
 	lists.fetch({
 		success: function(resp) {
 			itemData = {resp.toJSON()};
+			ReactDOM.render(<Sub itemData={itemData} />, document.getElementById('itemContainer'));
 			console.log('success ', resp);
 		},
 		error: function(err) {
@@ -32,7 +32,7 @@ var ListItem = Backbone.Model.extend({
 		},
 	});
 
-var Item = React.createClass({
+var Sub = React.createClass({
 	render: function() {
 		var listData = this.props.itemData.map(function(obj){
 			return (
@@ -45,4 +45,40 @@ var Item = React.createClass({
 	}
 });
 
-module.exports = listItem;
+var ToDo = React.createClass({
+	_handleSubmit: function(e) {
+		e.preventDefault();
+		var newItem = new ListItem();
+
+		newItem.set({
+			title: $("#newToDo").val()
+		})
+		title: $('#newToDo').val("");
+			newItem.save(null, {
+				success: function(resp) {
+					console.log('success ', resp);
+					listCollection.fetch({
+						success: function(resp){
+						},
+						error: function(err){
+						}
+					})
+				},
+				error: function(err) {
+					console.log('err ');
+				} 
+			})
+	},
+	render: function(){
+		return (
+			<div>
+				<h1>To Do App</h1>
+				<form onSubmit={this._handleSubmit}>
+					<input id="newToDo" placeholder="What needs to be done?"/>
+				</form>
+			</div>
+		)
+	}
+});
+
+ReactDOM.render(<ToDo />, document.getElementById('mainContainer'));
