--- js/components/mainApp.jsx
@@ -1,8 +1,0 @@
-var React = require ('react');
-var ReactDOM = require ('react-dom');
-var Backbone = require ('backbone');
-var BackboneParse = require ('../../backbone-parse.js');
-var Todo = require ('./listToDo.jsx');
-var listItem = ('./listItem.jsx');
-
-
