+++ js/main.jsx
@@ -1,6 +1,4 @@
 require ('../css/main.css');
 var React = require ('react');
 var ReactDOM = require ('react-dom');
-var App = require ('./components/mainApp.jsx');
-var listItem = ('./components/listItem.jsx');
-var Todo = require ('./components/listToDo.jsx');
+var ListItem = ('./components/listItem.jsx');
